requirements.txt
soccer-twos
torch>=1.9.0
PyYAML>=5.4
matplotlib>=3.3
aiohttp==3.7.4
aioredis==1.3.1
dm-tree==0.1.6
gym==0.19.0
gym-unity==0.27.0
mlagents==0.27.0
mlagents-envs==0.27.0
numpy==1.19.5
ray[rllib,tune]==1.4.0
